Throttle the rate of renewal requests (ACME v2 rate limit)
The ACME v2 API has a new rate limit of 300 new orders per account per 3 hours, which easily results in the "too many new orders recently" error for new or renewing domains. (see https://github.com/GUI/lua-resty-auto-ssl/issues/213)

As a first step to mitigate this new requirement, throttle the rate with which renewal requests are issued to a default 60 per hour (which leaves 40 for on-demand issues per hour).