Track Let's Encrypt certificate failures per-domain (ACME v2 rate limit)
This implements the minimal functionality for considering a domain's past certificate failures inside allow_domain. We believe something like this to be necessary to be able to comply to the new ACME v2 limit of 300 new orders/account/3h.

A different approach which might be worth considering is saving the statistical information to (permanent) storage, like we do with certificates. This patch instead only uses a shm-based dictionary, meaning that information might be stale in a multi-server setup. (Dictionary entries are created with an expiration date, and we're running the following patch to look for a certificate in storage before calling allow_domain, meaning that this shouldn't be a problem in practice: https://github.com/Cargo/lua-resty-auto-ssl/commit/b1f9715dd24a4ad4c1204d638fac10b2d0d23830)